web与controller

+++ src/main/java/scc1/java_project/deliver_server/POJO/Customer.java
@@ -3,8 +3,9 @@ package scc1.java_project.deliver_server.POJO;
 import lombok.Data;
 
 @Data
+// 客户
 public class Customer {
-    private int customerId;
+    private long customerId;
     private long customerPhoneNumber;
     private String customerPassword;
     private String customerName;
